refactor: move streamx into pkg dir
