+++ pkg/remote/trans/streamx/server_handler.go
@@ -28,7 +28,7 @@ import (
 	"github.com/cloudwego/kitex/pkg/klog"
 	"github.com/cloudwego/kitex/pkg/remote"
 	"github.com/cloudwego/kitex/pkg/rpcinfo"
-	"github.com/cloudwego/kitex/streamx"
+	"github.com/cloudwego/kitex/pkg/streamx"
 )
 
 /* 实际上 remote.ServerTransHandler 真正被 trans_server.go 使用的接口只有：
