+++ pkg/streamx/provider/jsonrpc/client_option.go
