+++ pkg/streamx/provider/jsonrpc/client_provier.go
@@ -1,24 +1,19 @@
-package ttstream
+package jsonrpc
 
 import (
 	"context"
-	"log"
-	"runtime"
-
-	"github.com/bytedance/gopkg/cloud/metainfo"
-	"github.com/cloudwego/gopkg/protocol/ttheader"
 	"github.com/cloudwego/kitex/client/streamxclient/streamxcallopt"
 	"github.com/cloudwego/kitex/pkg/kerrors"
 	"github.com/cloudwego/kitex/pkg/rpcinfo"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
-	"github.com/cloudwego/kitex/streamx"
+	"github.com/cloudwego/kitex/pkg/streamx"
+	"net"
 )
 
 var _ streamx.ClientProvider = (*clientProvider)(nil)
 
 func NewClientProvider(sinfo *serviceinfo.ServiceInfo, opts ...ClientProviderOption) (streamx.ClientProvider, error) {
 	cp := new(clientProvider)
-	cp.transPool = newTransPool(sinfo)
 	cp.sinfo = sinfo
 	for _, opt := range opts {
 		opt(cp)
@@ -27,7 +22,6 @@ func NewClientProvider(sinfo *serviceinfo.ServiceInfo, opts ...ClientProviderOpt
 }
 
 type clientProvider struct {
-	transPool    *transPool
 	sinfo        *serviceinfo.ServiceInfo
 	payloadLimit int
 }
@@ -39,27 +33,15 @@ func (c clientProvider) NewStream(ctx context.Context, ri rpcinfo.RPCInfo, callO
 	if addr == nil {
 		return nil, kerrors.ErrNoDestAddress
 	}
-
-	trans, err := c.transPool.Get(addr.Network(), addr.String())
+	conn, err := net.Dial(addr.Network(), addr.String())
 	if err != nil {
 		return nil, err
 	}
-
-	header := map[string]string{
-		ttheader.HeaderIDLServiceName: c.sinfo.ServiceName,
-	}
-	metainfo.SaveMetaInfoToMap(ctx, header)
-	s, err := trans.newStream(ctx, method, header)
+	trans := newTransport(c.sinfo, conn)
+	s, err := trans.newStream(method)
 	if err != nil {
 		return nil, err
 	}
 	cs := newClientStream(s)
-	runtime.SetFinalizer(cs, func(cs *clientStream) {
-		log.Printf("client stream[%v] closing", cs.sid)
-		_ = cs.close()
-		// TODO: current using one conn one stream
-		//_ = trans.close()
-		c.transPool.Put(trans)
-	})
 	return cs, err
 }
