+++ pkg/streamx/provider/jsonrpc/jsonrpc_impl_test.go
