+++ pkg/streamx/provider/jsonrpc/metadata.go
