+++ pkg/streamx/provider/jsonrpc/server_option.go
