+++ pkg/streamx/provider/jsonrpc/server_provider.go
@@ -1,13 +1,11 @@
-package ttstream
+package jsonrpc
 
 import (
 	"context"
 	"net"
 
-	"github.com/bytedance/gopkg/cloud/metainfo"
-	"github.com/cloudwego/gopkg/protocol/ttheader"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
-	"github.com/cloudwego/kitex/streamx"
+	"github.com/cloudwego/kitex/pkg/streamx"
 	"github.com/cloudwego/netpoll"
 )
 
@@ -30,28 +28,16 @@ type serverProvider struct {
 }
 
 func (s serverProvider) Available(ctx context.Context, conn net.Conn) bool {
-	data, err := conn.(netpoll.Connection).Reader().Peek(8)
-	if err != nil {
-		return false
-	}
-	return ttheader.IsStreaming(data)
+	err := checkFrame(conn.(netpoll.Connection))
+	return err == nil
 }
 
 func (s serverProvider) OnActive(ctx context.Context, conn net.Conn) (context.Context, error) {
-	trans := newTransport(serverTransport, s.sinfo, conn.(netpoll.Connection))
+	trans := newTransport(s.sinfo, conn)
 	return context.WithValue(ctx, serverTransCtxKey{}, trans), nil
 }
 
 func (s serverProvider) OnInactive(ctx context.Context, conn net.Conn) (context.Context, error) {
-	trans, _ := ctx.Value(serverTransCtxKey{}).(*transport)
-	if trans == nil {
-		return ctx, nil
-	}
-	// server should close transport
-	err := trans.close()
-	if err != nil {
-		return nil, err
-	}
 	return ctx, nil
 }
 
@@ -64,18 +50,12 @@ func (s serverProvider) OnStream(ctx context.Context, conn net.Conn) (context.Co
 	if err != nil {
 		return nil, nil, err
 	}
-	ctx = metainfo.SetMetaInfoFromMap(ctx, st.header)
 	ss := newServerStream(st)
 	return ctx, ss, nil
 }
 
 func (s serverProvider) OnStreamFinish(ctx context.Context, ss streamx.ServerStream) (context.Context, error) {
 	sst := ss.(*serverStream)
-	if err := sst.sendTrailer(); err != nil {
-		return nil, err
-	}
-	if err := sst.close(); err != nil {
-		return nil, err
-	}
-	return ctx, nil
+	err := sst.sendEOF()
+	return ctx, err
 }
