+++ pkg/streamx/provider/jsonrpc/transport.go
