+++ pkg/streamx/provider/ttstream/server_provider.go
@@ -1,11 +1,13 @@
-package jsonrpc
+package ttstream
 
 import (
 	"context"
 	"net"
 
+	"github.com/bytedance/gopkg/cloud/metainfo"
+	"github.com/cloudwego/gopkg/protocol/ttheader"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
-	"github.com/cloudwego/kitex/streamx"
+	"github.com/cloudwego/kitex/pkg/streamx"
 	"github.com/cloudwego/netpoll"
 )
 
@@ -28,16 +30,28 @@ type serverProvider struct {
 }
 
 func (s serverProvider) Available(ctx context.Context, conn net.Conn) bool {
-	err := checkFrame(conn.(netpoll.Connection))
-	return err == nil
+	data, err := conn.(netpoll.Connection).Reader().Peek(8)
+	if err != nil {
+		return false
+	}
+	return ttheader.IsStreaming(data)
 }
 
 func (s serverProvider) OnActive(ctx context.Context, conn net.Conn) (context.Context, error) {
-	trans := newTransport(s.sinfo, conn)
+	trans := newTransport(serverTransport, s.sinfo, conn.(netpoll.Connection))
 	return context.WithValue(ctx, serverTransCtxKey{}, trans), nil
 }
 
 func (s serverProvider) OnInactive(ctx context.Context, conn net.Conn) (context.Context, error) {
+	trans, _ := ctx.Value(serverTransCtxKey{}).(*transport)
+	if trans == nil {
+		return ctx, nil
+	}
+	// server should close transport
+	err := trans.close()
+	if err != nil {
+		return nil, err
+	}
 	return ctx, nil
 }
 
@@ -50,12 +64,18 @@ func (s serverProvider) OnStream(ctx context.Context, conn net.Conn) (context.Co
 	if err != nil {
 		return nil, nil, err
 	}
+	ctx = metainfo.SetMetaInfoFromMap(ctx, st.header)
 	ss := newServerStream(st)
 	return ctx, ss, nil
 }
 
 func (s serverProvider) OnStreamFinish(ctx context.Context, ss streamx.ServerStream) (context.Context, error) {
 	sst := ss.(*serverStream)
-	err := sst.sendEOF()
-	return ctx, err
+	if err := sst.sendTrailer(); err != nil {
+		return nil, err
+	}
+	if err := sst.close(); err != nil {
+		return nil, err
+	}
+	return ctx, nil
 }
