+++ pkg/streamx/provider/ttstream/transport_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/cloudwego/kitex/internal/test"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
-	"github.com/cloudwego/kitex/streamx"
+	"github.com/cloudwego/kitex/pkg/streamx"
 	"github.com/cloudwego/netpoll"
 )
